Refactored perspective.parallel.js to support multiple messaging implementations
