+++ packages/perspective/src/js/perspective.parallel.js
@@ -11,6 +11,8 @@ import {detectIE, ScriptPath} from "./utils.js";
 
 import {TYPE_AGGREGATES, AGGREGATE_DEFAULTS, TYPE_FILTERS, FILTER_DEFAULTS, SORT_ORDERS} from "./defaults.js";
 
+import {worker} from "./api.js";
+
 /******************************************************************************
  *
  * Utilities
@@ -31,188 +33,6 @@ if (detectIE() && window.location.href.indexOf(__SCRIPT_PATH__.host()) === -1) {
     }(document));
 }
 
-/******************************************************************************
- *
- * API
- *
- */
-
-function subscribe(method, cmd) {
-    return function() {
-        var resolve = arguments[arguments.length - 1];
-        var args = Array.prototype.slice.call(arguments, 0, arguments.length - 1);
-        this._worker.handlers[++this._worker.msg_id] = {resolve, reject: () => {}, keep_alive: true};
-        var msg = {
-            id: this._worker.msg_id,
-            cmd: cmd || 'view_method',
-            name: this._name,
-            method: method,
-            args: args,
-            subscribe: true
-        };
-        if (this._worker.initialized.value) {
-            this._worker.postMessage(msg);
-        } else {
-            this._worker.messages.push(() => this._worker.postMessage(msg));
-        }
-    }
-}
-
-function async_queue(method, cmd) {
-    return function() {
-        var args = Array.prototype.slice.call(arguments, 0, arguments.length);
-        return new Promise(function(resolve, reject) {
-            this._worker.handlers[++this._worker.msg_id] = {resolve, reject};
-            var msg = {
-                id: this._worker.msg_id,
-                cmd: cmd || 'view_method',
-                name: this._name,
-                method: method,
-                args: args,
-                subscribe: false
-            };
-            if (this._worker.initialized.value) {
-                this._worker.postMessage(msg);
-            } else {
-                this._worker.messages.push(() => this._worker.postMessage(msg));
-            }
-        }.bind(this));
-    };
-}
-
-function view(table_name, worker, config) {
-    this._worker = worker;
-    this._config = config;
-    this._name = Math.random() + "";
-    var msg = {
-        cmd: 'view',
-        view_name: this._name,
-        table_name: table_name,
-        config: config
-    }
-    if (this._worker.initialized.value) {
-        this._worker.postMessage(msg);
-    } else {
-        this._worker.messages.push(() => this._worker.postMessage(msg));
-    }
-}
-
-view.prototype.to_json = async_queue('to_json');
-
-view.prototype.to_csv = async_queue('to_csv');
-
-view.prototype.schema = async_queue('schema');
-
-view.prototype.num_columns = async_queue('num_columns');
-
-view.prototype.num_rows = async_queue('num_rows');
-
-view.prototype.expand_to_depth = async_queue('expand_to_depth');
-
-view.prototype.collapse_to_depth = async_queue('collapse_to_depth');
-
-view.prototype.get_row_expanded = async_queue('get_row_expanded');
-
-view.prototype.open = async_queue('open');
-
-view.prototype.close = async_queue('close');
-
-view.prototype.delete = async_queue('delete');
-
-view.prototype.on_update = subscribe('on_update', 'view_method', true);
-
-view.prototype.on_delete = subscribe('on_delete', 'view_method', true);
-
-
-function table(worker, name) {
-    this._name = name;
-    this._worker = worker;
-}
-
-table.prototype.add_computed = function(computed) {
-    let name = Math.random() + "";
-    let original = this._name;
-    // serialize functions
-    for (let i = 0; i < computed.length; ++i) {
-        let column = computed[i];
-        let func = column['func'];
-        if (typeof func == "function") {
-            column['func'] = func.toString();
-        }
-    }
-    var msg = {
-        cmd: 'add_computed',
-        original: original,
-        name: name,
-        computed: computed
-    };
-    if (this._worker.initialized.value) {
-        this._worker.postMessage(msg);
-    } else {
-        this._worker.messages.push(() => this._worker.postMessage(msg));
-    }
-    return new table(this._worker, name);
-}
-
-table.prototype.worker = function () {
-    return this._worker;
-}
-
-table.prototype.view = function (config) {
-    return new view(this._name, this._worker, config);
-}
-
-table.prototype.schema = async_queue('schema', 'table_method');
-
-table.prototype.size = async_queue('size', 'table_method');
-
-table.prototype.columns = async_queue('columns', 'table_method');
-
-table.prototype.delete = async_queue('delete', "table_method");
-
-table.prototype.on_delete = subscribe('on_delete', 'table_method', true);
-
-table.prototype.remove = async_queue('remove', "table_method");
-
-table.prototype.update = function(data) {
-    return new Promise( (resolve, reject) => {
-        this._worker.handlers[++this._worker.msg_id] = {resolve, reject};
-        var msg = {
-            id: this._worker.msg_id,
-            name: this._name,
-            cmd: 'table_method',
-            method: 'update',
-            args: [data],
-        };
-        let post = () => {
-            if (this._worker.transferable && data instanceof ArrayBuffer) {
-                this._worker.postMessage(msg, [data]);
-            } else {
-                this._worker.postMessage(msg);
-            }
-        };
-        if (this._worker.initialized.value) {
-            post();
-        } else {
-            this._worker.messages.push(post);
-        }
-    });
-}
-
-
-table.prototype.execute = function (f) {
-    var msg = {
-        cmd: 'table_execute',
-        name: this._name,
-        f: f.toString()
-    }
-    if (this._worker.initialized.value) {
-        this._worker.postMessage(msg);
-    } else {
-        this._worker.messages.push(() => this._worker.postMessage(msg));
-    }
-}
-
 function XHRWorker(url, ready, scope) {
     var oReq = new XMLHttpRequest();
     oReq.addEventListener('load', function() {
@@ -227,20 +47,26 @@ function XHRWorker(url, ready, scope) {
     oReq.send();
 }
 
-function worker() {
-    this._worker = {
-        initialized: {value: false},
-        transferable: false,
-        msg_id: 0,
-        handlers: {},
-        messages: []
-    };
-    if (window.__PSP_WORKER__) {
-        this._start_embedded();
-    } else if (window.location.href.indexOf(__SCRIPT_PATH__.host()) > -1) {
-        this._start_same_origin();
+class WebWorker extends worker {
+
+    constructor() {
+        super();
+        if (window.__PSP_WORKER__) {
+            this._start_embedded();
+        } else if (window.location.href.indexOf(__SCRIPT_PATH__.host()) > -1) {
+            this._start_same_origin();
+        } else {
+            this._start_cross_origin();
+        }
+    
+    }
+}
+
+WebWorker.prototype.send = function (msg) {
+    if (this._worker.transferable && msg.args && msg.args[0] instanceof ArrayBuffer) {
+        this._worker.postMessage(msg, msg.args);
     } else {
-        this._start_cross_origin();
+        this._worker.postMessage(msg);
     }
 }
 
@@ -313,11 +139,11 @@ worker.prototype._start_cross_origin_wasm = function () {
 }
 
 // https://github.com/kripken/emscripten/issues/6042
-function detect_iphone() {
+function detect_iphone () {
     return /iPad|iPhone|iPod/.test(navigator.userAgent) && !window.MSStream;
 }
 
-worker.prototype._start_same_origin = function() {
+worker.prototype._start_same_origin = function () {
     var dir = (typeof WebAssembly === "undefined" || detect_iphone() ? 'asmjs' : 'wasm_async');
     var w =  new Worker(__SCRIPT_PATH__.path() + dir + '/perspective.js');
     for (var key in this._worker) {
@@ -329,74 +155,17 @@ worker.prototype._start_same_origin = function() {
     this._detect_transferable();
 };
 
-let _initialized = false;
-
-worker.prototype._handle = function(e) {
-    if (!this._worker.initialized.value) {
-        if (!_initialized) {
-            var event = document.createEvent("Event");
-            event.initEvent("perspective-ready", false, true);
-            window.dispatchEvent(event);
-            _initialized = true;
-        }
-        for (var m in this._worker.messages) {
-            if (this._worker.messages.hasOwnProperty(m)) {
-                this._worker.messages[m]();
-            }
-        }
-        this._worker.initialized.value = true;
-        this._worker.messages = [];
-    }
-    if (e.data.id) {
-        var handler = this._worker.handlers[e.data.id];
-        if (handler) {
-            if (e.data.error) {
-                handler.reject(e.data.error);
-            } else {
-                handler.resolve(e.data.data);
-            }
-            if (!handler.keep_alive) {
-                delete this._worker.handlers[e.data.id];
-            }
-        }
-    }
-};
-
-worker.prototype.table = function(data, options) {
-    // Set up msg
-    name = Math.random() + "";
-    var msg = {
-        cmd: 'table',
-        name: name,
-        data: data,
-        options: options || {}
-    };
-    let post = () => {
-        if (this._worker.transferable && data instanceof ArrayBuffer) {
-            this._worker.postMessage(msg, [data]);
-        } else {
-            this._worker.postMessage(msg);
-        }
-    };
-    if (this._worker.initialized.value) {
-        post();
-    } else {
-        this._worker.messages.push(post);
-    }
-    return new table(this._worker, name);
-};
-
-worker.prototype.terminate = function() {
+worker.prototype.terminate = function () {
     this._worker.terminate();
     this._worker = undefined;
 };
 
 export default {
-    worker: function() {
-        if (window.location.href.indexOf(__SCRIPT_PATH__.host()) === -1 && detectIE()) {
-            return perspective;
-        }
-        return new worker();
+    worker: function () {
+        // if (window.location.href.indexOf(__SCRIPT_PATH__.host()) === -1 && detectIE()) {
+        //     return perspective;
+        // }
+        return new WebWorker();
     },
 
     TYPE_AGGREGATES: TYPE_AGGREGATES,
